libnm,core: add "${UUID}" stable ID to explicitly select default
The "connection.stable-id" supports placeholders like "${CONNECTION}" or
"${DEVICE}".

The stable-id can also be specified in global connection defaults in
NetworkManager.conf, by leaving it unset in the profile. Global
connection defaults always follow the pattern, that they correspond to a
per-profile property, and only when the per-profile value indicates a
special default/unset value, the global connection default is consulted.

One requirement is that when a fallback to the global connection default
is used, that we can configure thereby a value that behaves the same as if
we:
 - either would specify the value per-profile
 - or would not set it per profile nor NetworkManager.conf, but fallback
   to the built-in default.

The built-in default of the stable-id is generated based on
"connection.uuid", but that value was not configurable explicitly,
neither per-profile nor in NetworkManager.conf.

For example compare to ipv6.ip6-privacy property which supports the
values "disabled", "prefer-public-addr", "prefer-temp-addr" and
"unknown". If set the per-profile value to "unknown", a global
connection default is consulted. If not set in NetworkManager.conf, the
build-in default depends on the
/proc/sys/net/ipv6/conf/default/use_tempaddr. But in every case, the
used value is one of "disabled", "prefer-public-addr", or
"prefer-temp-addr", and you could explicitly set that value either in
the profile or in NetworkManager.conf.

Fix that by introducing "${UUID}" to represent the built-in default.
"{$UUID}" in combination with other strings (like "${UUID}foo") behaves
similar to "${CONNECTION}foo" -- but does not yield exactly the same
value, because the strings "${UUID}foo" and "${CONNECTION}foo" are
different.