core: rate-limit updates to IP addresses/routes on D-Bus API
It doesn't scale. If you add 100k routes one-by-one, then upon each
change from platform, we will send the growing list of the routes on
D-Bus.

That is too expensive. Especially, if you imagine that the receiving end
is a NMClient instance. There is a D-Bus worker thread that queues the
received GVariant messages, while the main thread may not be able to
process them fast enough. In that case, the memory keeps growing very
fast and due to fragmentation it is not freed.

Instead, rate limit updates to 3 per second.

Note that the receive buffer of the netlink socket can fill up and we
loose messages. Therefore, already on the lowest level, we may miss
addresses/routes. Next, on top of NMPlatform, NMIPConfig listens to
NM_L3_CONFIG_NOTIFY_TYPE_PLATFORM_CHANGE_ON_IDLE events. Thereby it
further will miss intermediate state (e.g. a route that exists only for
a short moment).

Adding another delay and rate limiting on top of that, does not
fundamentally change the fact, that we never see all events from
netlink. We may miss addresses/routes that only exist for a shit amount
of time. This makes "the problem" worse, but not fundamentally new.

We can only get a (correct) settled state, after all events are
processed (and we don't really know, whether there is an event just
about to be received).

Rate limiting is important to not overwhelm D-Bus clients. In reality,
none of the users really need this information, because it's also
incomplete. Users who really need to know the D-Bus should use netlink
or find another way (a way that scales and where the explicitly request
this information).