cli: fix setting "wifi.mac-address-randomization"
With `nmcli connection modify`, later options should overwrite earlier
ones. That did not work correctly with

  nmcli --offline connection add type wifi \
        wifi.ssid xxxx \
        wifi.cloned-mac-address permanent \
        wifi.mac-address-randomization 0

That's because "wifi.mac-address-randomization" is a mostly redundant
alias for certain "wifi.cloned-mac-address" options, and libnm does
various normalizations to make that somewhat seamless.

However, once "cloned-mac-address" property is set, setting any value of
"wifi.mac-address-randomization" has no effect, as it gets normalized
away by libnm. This is a sensible thing to do, in most cases to best
handle the deprecation/aliasing.

For nmcli, if the user sets "wifi.mac-address-randomization", it really
means to also reset the "cloned-mac-address". Thus nmcli needs to do
extra work to get this right.