libnmc: get enum gtype for enum properties implemented as int
In nm-meta-setting-desc we were getting the enum type from the
GSpecParam for those properties implemented as enum. For properties
implemented as int but exposed to nmcli as enum, we needed to set
get_gtype to define the enum type.

Properties defined as direct_enum are implemented as integer but have an
associated enum type. Use it in nm-meta-setting-desc.